login refactor, business logic moved to service, tested
